requirement.txt
uvicorn
fastapi
requests 
python-dotenv
sqlalchemy
psycopg2-binary
passlib
bcrypt==4.0.1
python-multipart
pyjwt
pwdlib[argon2]